Downsizing of data types for reasons of memory usage and also speed

+++ src/at/sume/db/RecordSet.java
@@ -162,7 +162,7 @@ public abstract class RecordSet<E extends RecordSetRow<?>> implements Iterable<E
 	 * @param id
 	 * @return
 	 */
-	public E lookup(Long id) {
+	public E lookup(Integer id) {
 		E lookupKey = createRecordSetRow();
 		lookupKey.setId(id);
 		int i = Collections.binarySearch(rowList, lookupKey);
